<==== Mercurial data ====>
Branch: vms_3.1_dev

+++ unit_tests/nx_network_ut/src/http/server/server_connection_ut.cpp
@@ -243,7 +243,7 @@ TEST_F(HttpAsyncServerConnectionTest, multipleRequestsTest)
         SocketFactory::tcpServerIpVersion());
 
     ASSERT_TRUE(socket->connect(m_testHttpServer->serverAddress()));
-    ASSERT_EQ(sizeof(testData) - 1, socket->send(testData, sizeof(testData) - 1));
+    ASSERT_EQ((int)sizeof(testData) - 1, socket->send(testData, sizeof(testData) - 1));
 }
 
 TEST_F(HttpAsyncServerConnectionTest, inactivityTimeout)
